Refactor User enum fields into enums.py
