+++ itou/france_connect/tests.py
@@ -5,6 +5,7 @@ from django.test import TestCase
 from django.urls import reverse
 from django.utils import timezone
 
+from ..users.enums import IdentityProvider
 from ..users.factories import UserFactory
 from ..users.models import User
 from . import models as france_connect_models, views as france_connect_views
@@ -74,14 +75,14 @@ class FranceConnectTest(TestCase):
         self.assertEqual(user.last_name, user_data["family_name"])
         self.assertEqual(user.first_name, user_data["given_name"])
         self.assertEqual(user.external_data_source_history["last_name"]["source"], "franceconnect")
-        self.assertEqual(user.identity_provider, User.IdentityProvider.FRANCE_CONNECT)
+        self.assertEqual(user.identity_provider, IdentityProvider.FRANCE_CONNECT)
 
         # Update user
         fc_user_data.last_name = "DUPUIS"
         user, created = france_connect_models.create_or_update_user(fc_user_data)
         self.assertFalse(created)
         self.assertEqual(user.last_name, "DUPUIS")
-        self.assertEqual(user.identity_provider, User.IdentityProvider.FRANCE_CONNECT)
+        self.assertEqual(user.identity_provider, IdentityProvider.FRANCE_CONNECT)
 
     def test_create_user_from_user_data_with_already_existing_fc_id(self):
         """
@@ -96,7 +97,7 @@ class FranceConnectTest(TestCase):
         self.assertEqual(user.last_name, user_data["family_name"])
         self.assertEqual(user.first_name, user_data["given_name"])
         self.assertEqual(user.external_data_source_history["last_name"]["source"], "franceconnect")
-        self.assertEqual(user.identity_provider, User.IdentityProvider.FRANCE_CONNECT)
+        self.assertEqual(user.identity_provider, IdentityProvider.FRANCE_CONNECT)
 
     def test_create_user_from_user_data_with_already_existing_fc_email(self):
         """
@@ -112,7 +113,7 @@ class FranceConnectTest(TestCase):
         self.assertNotEqual(user.first_name, user_data["given_name"])
         # We did not fill this data using external data, so it is not set
         self.assertIsNone(user.external_data_source_history)
-        self.assertNotEqual(user.identity_provider, User.IdentityProvider.FRANCE_CONNECT)
+        self.assertNotEqual(user.identity_provider, IdentityProvider.FRANCE_CONNECT)
 
     def test_callback_no_code(self):
         url = reverse("france_connect:callback")
